Update sonata_tech.rst
Defining the third cell class of synapses which is modulatory (In my case: cholinergic interneurons) which can be excitatory or inhibitory depends upon postsynaptic neuron.
There are three types of cells in the striatum, excitatory, inhibitory, and neuromodulatory. In building the striatum circuit, we need to define a third class of cells that are neuromodulatory and whose synapse class depends upon their postsynaptic neurons.